Adds support for checking for commitment before a trjaectory has been initialized in Amber

+++ atesa_v2/jobtype.py
@@ -296,7 +296,7 @@ class AimlessShooting(JobType):
             for job_index in range(len(self.jobids)):
                 if self.get_status(job_index, settings) == 'R':     # if the job in question is running
                     frame_to_check = self.get_frame(self.history.prod_trajs[-1][job_index], -1, settings)
-                    if utilities.check_commit(frame_to_check, settings):  # if it has committed to a basin
+                    if frame_to_check and utilities.check_commit(frame_to_check, settings):  # if it has committed to a basin
                         self.cancel_job(job_index, settings)        # cancel it
                         os.remove(frame_to_check)
